requirements.txt
python-dotenv
qdrant-client
datasets
fastembed
pillow
torch
google-genai
fastapi
uvicorn
requests

# Core dependencies for ad generation agent
qdrant-client>=1.9.0
fastembed>=0.3.5
google-genai>=1.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.32.0

# API server dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data processing (for ingestion scripts)
datasets>=3.0.0
tqdm>=4.65.0

# Optional: for demo UI
streamlit>=1.28.0
